# Switched to DataDog zstd wrapper, reusing the compression ctx